pyproject.toml
[project]
name = "chzzk-data-analytics"
version = "0.1.0"
description = ""
authors = [
    {name = "uowol",email = "alllhvhla@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pyaml (>=25.1.0,<26.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "duckdb (>=1.3.0,<2.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "fastparquet (>=2024.11.0,<2025.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "selenium (>=4.33.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "kafka-python (>=2.2.13,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
